// src/applications/domain/application-event-type.enum.ts
export enum ApplicationEventType {
  APPLICATION_CREATED = 'APPLICATION_CREATED',
  APPLICATION_UPDATED = 'APPLICATION_UPDATED',
  OFFER_SIMULATION_SUCCEEDED = 'OFFER_SIMULATION_SUCCEEDED',
  OFFER_SIMULATION_NOT_VIABLE = 'OFFER_SIMULATION_NOT_VIABLE',
  OFFER_SIMULATION_TECHNICAL_ERROR = 'OFFER_SIMULATION_TECHNICAL_ERROR',
  APPLICATION_FINALIZED = 'APPLICATION_FINALIZED',
  APPLICATION_ABANDONED = 'APPLICATION_ABANDONED',
  STATUS_CHANGED = 'STATUS_CHANGED',
}